# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for G4 Car Service Drivers Application.",
        "title": "G4 Drivers API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "samllachp@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/drivers/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit driver application with documents and photos",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Register as Driver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Full Name",
                        "name": "full_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Phone Number",
                        "name": "phone_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Emergency Contact",
                        "name": "emergency_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device Type (phone/tablet)",
                        "name": "device_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Vehicle Type",
                        "name": "vehicle_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Passenger Capacity",
                        "name": "passenger_capacity",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Category (comfort/luxury)",
                        "name": "driver_category",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Driver's License",
                        "name": "driver_license",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "TLC License",
                        "name": "tlc_license",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Car Registration",
                        "name": "car_registration",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Vehicle Inspection",
                        "name": "vehicle_inspection",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "TLC Diamond",
                        "name": "tlc_diamond",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Profile Photo",
                        "name": "profile_photo",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Vehicle Photos (Array)",
                        "name": "vehicle_photos",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Insurance Files (Array)",
                        "name": "insurance_files",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JSON String with extras",
                        "name": "additional_info",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "{\"status\":\"success\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}